% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_interpolation.R
\docType{class}
\name{configParser-class}
\alias{configParser-class}
\alias{ConfigParser}
\title{R6 object to parse an INI file}
\description{
This class creates an object that represents an INI file. It
also supports interpolation of variables of the form \code{\%(...)s}.
More detail on how the interpolation works in given in the
explanation of method \code{get}.
}
\section{Methods}{

\describe{
\item{\code{new(init = NULL)}}{

 Initializes the object

Initializes the object, optionally with pre-set variables given
as a list or an environment

\strong{Parameters:}
\describe{
\item{init}{A list or an environment of variables to pre-set (the will be
put into the \code{DEFAULT} section}
}}

\item{\code{read(filepath, encoding = getOption("encoding"))}}{

Read in an INI file

Reads the INI given into the object. It will be added to the
internal storage of the object. If applying these functions
several times, later definitions overwrite those that
were read earlier.



\strong{Parameters:}
\describe{
\item{filepath}{The path to the INI file.}
\item{encoding}{The encoding to use}
}

\strong{Return:}

The object, but with the added data of the file


}

\item{\code{write(filepath, encoding = getOption("encoding"))}}{

Write an INI file

Write the ConfigParser object into the INI file. It writes all the different sections
using uninterpolated variables.



\strong{Parameters:}
\describe{
\item{filepath}{The path to the INI file.}
\item{encoding}{The encoding to use}
}

\strong{Return:}

The ConfigParser object itself


}

\item{\code{get(key, fallback, section = "default", interpolate = TRUE)}}{

Get the value of a key



\strong{Parameters:}
\describe{
\item{key}{The key for which to get the value}
\item{fallback}{The fallback value to return if there is no value for the key. If missing,
an error will be thrown if the key has no value.}
\item{section}{The section (or several) from which to read the key. It will try to read the
key in the given section from first to last, with \code{DEFAULT} always being the
last}
\item{interpolate}{Should the values be interpolated. This will try to replace variables of
the form \code{\%(...)s}}
}

\strong{Return:}

The value of the key


}

\item{\code{getboolean(key, fallback, section = "default", interpolate = TRUE)}}{

Same as \code{get}, but results coerced to a logical.}

\item{\code{getfloat(key, fallback, section = "default", interpolate = TRUE)}}{

Same as \code{get}, but the result coerced to a float.}
}}

