% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_interpolation.R
\name{interpolation_key}
\alias{interpolation_key}
\alias{do_replacement}
\title{Search for key's for interpolation}
\usage{
interpolation_key(value)

do_replacement(input, key, replacement)
}
\arguments{
\item{value}{The value to search for possibly required interpolation}
}
\value{
A string representing the key of NULL
}
\description{
Search for key's for interpolation
}
\details{
Searches the string for occurences of %(...)s that needs interpolation
and return the unchanged value of ... as the key;

This can then later be used with \code{do_replacement} and a value
to perform the actual replacement.
}
\author{
Holger Hoefling
}
\keyword{internal}
