% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{interpolation_option}
\alias{interpolation_option}
\alias{do_replacement}
\title{Search for option's for interpolation}
\usage{
interpolation_option(value)

do_replacement(input, option, replacement)
}
\arguments{
\item{value}{The value to search for possibly required interpolation}
}
\value{
A string representing the option of NULL
}
\description{
Search for option's for interpolation
}
\details{
Searches the string for occurences of %(...)s that needs interpolation
and return the unchanged value of ... as the option;

This can then later be used with \code{do_replacement} and a value
to perform the actual replacement.
}
\author{
Holger Hoefling
}
\keyword{internal}
